Move stuff to where they should be
